LP1901760 Improve SharedWorker non-support handling (AngJS)
Force the SharedWorker code to run before org settings cache requests
are attempted so that we can tell before the lovefield request is made
wether or not a connection will be successfull.  In other words, we may
not know if cannotConnect is true yet, so force that logic to run first.

Signed-off-by: Bill Erickson <berickxx@gmail.com>
Signed-off-by: Terran McCanna <tmccanna@georgialibraries.org>
Signed-off-by: Llewellyn Marshall <llewellyn.marshall@ncdcr.gov>